+++ docs/CHANGELOG.md
@@ -51,6 +51,7 @@
     * `Render Vignette` (Enabled By Default)
     * `Increase Render Chunk Size` (Enabled By Default)
     * `Proper Entity Shading` (Enabled By Default)
+    * `Fix Sugar Position In Hand` (Enabled By Default)
   * Existing Functionality (All Enabled By Default)
     * `Fix Screen Rendering When Hiding HUD`
     * `Sanitize Usernames`